Added objectEncoding patch from Prabhu
git-svn-id: http://red5.googlecode.com/svn/java/server/trunk@3322 1b6495e4-3631-0410-8e05-8f51eee8b9cc

+++ src/org/red5/server/net/rtmp/BaseRTMPClientHandler.java
@@ -271,7 +271,7 @@ public abstract class BaseRTMPClientHandler extends BaseRTMPHandler {
 		this.connectArguments = connectCallArguments;
 
 		if (!connectionParams.containsKey("objectEncoding")) {
-			connectionParams.put("objectEncoding", (double) 0);
+			connectionParams.put("objectEncoding", (int) 0);
 		}
 
 		this.connectCallback = connectCallback;
